pyproject.toml
[tool.poetry]
name = "symnet-cp"
version = "0.5.0"
description = "SymNet External Control Protocol implementation"
authors = ["Christian Kohlstedde <christian@kohlsted.de>"]
license = "MIT"
readme = "README.md"

keywords = ["Symetrix", "SymNet", "Solus"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Natural Language :: German",
]

[tool.poetry.dependencies]
python = "^3.11"
attrs = "^24.1.0"
prometheus-client = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.2"
black = "^24.8.0"
pytest-aio = "^1.9.0"
pytest-mockservers = "^0.6.0"
pytest-cov = "^5.0.0"
flake8 = "^7.1.1"
pytest-sugar = "^1.0.0"
isort = "^5.10.1"
pep8-naming = "^0.14.1"
tomli = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py311']

[tool.isort]
profile = "black"
skip_gitignore = true

[tool.coverage.run]
branch = true
